refactor: cleanup routes, functions and update packages
