+++ modules/api/package.json
@@ -49,21 +49,15 @@
     "@types/express": "^5.0.6",
     "@types/jest": "^30.0.0",
     "@types/js-yaml": "^4.0.9",
-    "@types/node": "^20.19.27",
     "@types/supertest": "^6.0.3",
-    "eslint": "^9.39.1",
-    "globals": "^15.15.0",
     "jest": "^30.2.0",
     "js-yaml": "^4.1.1",
-    "prettier": "^3.7.4",
     "source-map-support": "^0.5.21",
     "supertest": "^7.1.4",
     "ts-jest": "^29.4.6",
     "ts-loader": "^9.5.4",
     "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.9.3",
-    "typescript-eslint": "^8.50.1"
+    "tsconfig-paths": "^4.2.0"
   },
   "jest": {
     "moduleFileExtensions": [
