+++ modules/web/package.json
@@ -30,7 +30,7 @@
     "@radix-ui/react-tooltip": "^1.2.8",
     "@tanstack/react-form": "^1.27.7",
     "@tanstack/react-query": "^5.90.14",
-    "@tanstack/react-router": "^1.145.7",
+    "@tanstack/react-router": "^1.145.10",
     "@tanstack/react-start": "^1.145.10",
     "axios": "^1.13.2",
     "class-variance-authority": "^0.7.1",
@@ -62,25 +62,19 @@
     "@tailwindcss/postcss": "^4.1.18",
     "@tailwindcss/vite": "^4.1.18",
     "@types/negotiator": "^0.6.4",
-    "@types/node": "^20.19.27",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^5.1.2",
-    "eslint": "^9.17.0",
     "eslint-import-resolver-typescript": "^3.10.1",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-react-refresh": "^0.4.26",
     "eslint-plugin-sonarjs": "^3.0.5",
-    "globals": "^15.15.0",
     "orval": "^7.17.2",
-    "prettier": "^3.7.4",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "tailwindcss": "^4.1.18",
     "tw-animate-css": "^1.4.0",
-    "typescript": "^5.9.3",
-    "typescript-eslint": "^8.18.2",
     "vite": "^7.3.1",
     "vite-tsconfig-paths": "^6.0.3"
   }
