+++ modules/web/src/routes/locations.$id.tsx
