+++ modules/web/src/routes/stock.tsx
